Reorganized ProjectInformation and AnalysisController.
