+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/facade/BeagleController.java
@@ -4,7 +4,6 @@ import de.uka.ipd.sdq.beagle.core.AnalysisController;
 import de.uka.ipd.sdq.beagle.core.Blackboard;
 import de.uka.ipd.sdq.beagle.core.LaunchConfiguration;
 import de.uka.ipd.sdq.beagle.core.ProjectInformation;
-import de.uka.ipd.sdq.beagle.core.ProjectInformation.AnalysisState;
 import de.uka.ipd.sdq.beagle.core.analysis.MeasurementResultAnalyserContributionsHandler;
 import de.uka.ipd.sdq.beagle.core.analysis.ProposedExpressionAnalyserContributionsHandler;
 import de.uka.ipd.sdq.beagle.core.failurehandling.FailureHandler;
@@ -130,7 +129,7 @@ public class BeagleController {
 	 *
 	 */
 	public void pauseAnalysis() {
-		this.blackboard.getProjectInformation().setAnalysisState(AnalysisState.ENDING);
+		this.analysisController.setAnalysisState(AnalysisController.AnalysisState.ENDING);
 	}
 
 	/**
@@ -139,7 +138,7 @@ public class BeagleController {
 	 *
 	 */
 	public void continueAnalysis() {
-		this.blackboard.getProjectInformation().setAnalysisState(AnalysisState.RUNNING);
+		this.analysisController.setAnalysisState(AnalysisController.AnalysisState.RUNNING);
 	}
 
 	/**
@@ -149,7 +148,7 @@ public class BeagleController {
 	 *
 	 */
 	public void abortAnalysis() {
-		this.blackboard.getProjectInformation().setAnalysisState(AnalysisState.ABORTING);
+		this.analysisController.setAnalysisState(AnalysisController.AnalysisState.ABORTING);
 	}
 
 	/**
